+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+Breaking:
+
+ - Rename `feature` to `mixin`.
+
 ## v0.1.1
 
  - Allow importing factories from other modules ([#1](https://github.com/mjkillough/factori/pull/1) by @danbruder)